+++ src/core/prelude/i18n/const.ts
@@ -27,8 +27,7 @@ export const
  */
 export const locale: Locale = {
 	value: undefined,
-	isDefined: false,
-	isInitialized: Promise.resolve()
+	isDefault: false
 };
 
 /**