+++ src/core/prelude/i18n/index.ts
@@ -14,43 +14,27 @@
 import config from 'config';
 
 import { IS_NODE } from 'core/env';
-import type { AsyncStorageNamespace } from 'core/kv-storage';
 
 import { emitter, locale } from 'core/prelude/i18n/const';
+
+import storage from 'core/prelude/i18n/storage';
 
 export * from 'core/prelude/i18n/const';
 export * from 'core/prelude/i18n/helpers';
 export * from 'core/prelude/i18n/interface';
 
-let
-	storage: CanUndef<Promise<AsyncStorageNamespace>>;
-
-//#if runtime has core/kv-storage
-// eslint-disable-next-line prefer-const
-storage = import('core/kv-storage').then(({asyncLocal}) => asyncLocal.namespace('[[I18N]]'));
-//#endif
-
 if (IS_NODE) {
 	setLocale(config.locale);
 
 } else {
-	locale.isInitialized = (async () => {
-		try {
-			const
-				s = await storage,
-				l = await s.get<Language>('locale');
+	const locale = storage.get<Language>('locale');
 
-			if (l != null) {
-				setLocale(l, await s.get<boolean>('isLocaleDef'));
-				return;
-			}
+	if (locale != null) {
+		setLocale(locale, storage.get<boolean>('isLocaleDef'));
 
-			throw new Error('Default language');
-
-		} catch {
-			setLocale(config.locale, true);
-		}
-	})();
+	} else {
+		setLocale(config.locale, true);
+	}
 }
 
 /**
@@ -69,13 +53,11 @@ export function setLocale(value: CanUndef<Language>, def?: boolean): CanUndef<La
 	}
 
 	locale.value = value;
-	locale.isDefined = Boolean(def);
+	locale.isDefault = Boolean(def);
 
-	if (!IS_NODE && storage != null) {
-		storage.then((storage) => Promise.all([
-			storage.set('locale', value),
-			storage.set('isLocaleDef', locale.isDefined)
-		])).catch(stderr);
+	if (!IS_NODE) {
+		storage.set('locale', value);
+		storage.set('isLocaleDef', locale.isDefault);
 	}
 
 	emitter.emit('setLocale', value, oldVal);